# server/openapi.yaml
openapi: 3.0.3
info:
  title: IDSC Portal - Faculty System API
  description: |
    API for the IDSC Portal Faculty System.

    The system allows faculty members to:
    - View their assigned subjects and sections
    - View students enrolled in their assigned subjects
    - Manage student grades
    - Manage student attendance
    - View schedules
    - View announcements

    Dean users have authorization to manage:
    - Faculty
    - Courses
    - Subjects
    - Sections
    - Faculty subject assignments
    - Schedules
    - Announcements
    - Student grades
    - Student attendance
  version: 1.4.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
...
paths:
...
  /faculty:
    get:
      tags:
        - Faculty
      summary: Get all faculty
      description: Dean can view all faculty members.
      x-roles:
        - DEAN
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Faculty list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FacultyItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Faculty
      summary: Create faculty
      description: Dean creates a new faculty member.
      x-roles:
        - DEAN
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacultyPayload'
      responses:
        '201':
          description: Faculty created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacultyItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /faculty/{facultyId}:
    parameters:
      - $ref: '#/components/parameters/FacultyId'

    get:
      tags:
        - Faculty
      summary: Get faculty by ID
      description: Get faculty information by faculty ID.
      x-roles:
        - DEAN
        - FACULTY
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Faculty retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacultyItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Faculty
      summary: Update faculty
      description: Dean updates faculty information.
      x-roles:
        - DEAN
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacultyPayload'
      responses:
        '200':
          description: Faculty updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacultyItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Faculty
      summary: Delete faculty
      description: Dean deletes a faculty member.
      x-roles:
        - DEAN
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Faculty deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==========================================================
  # PARAMETERS
  # ==========================================================
  parameters:
    FacultyId:
      name: facultyId
      in: path
      required: true
      description: Unique faculty identifier
      schema:
        type: string
      example: FAC-001
...
  schemas:
...
    FacultyItem:
      type: object
      required:
        - facultyId
        - employeeId
        - firstName
        - lastName
        - position
        - employmentType
        - status
      properties:
        facultyId:
          type: string
          example: FAC-001

        employeeId:
          type: string
          example: EMP-001

        firstName:
          type: string
          example: Juan

        lastName:
          type: string
          example: Dela Cruz

        middleName:
          type: string
          nullable: true
          example: Santos

        email:
          type: string
          format: email
          example: juan.delacruz@example.com

        position:
          type: string
          enum:
            - FACULTY
            - PROGRAM_HEAD
            - DEAN
          example: FACULTY

        employmentType:
          type: string
          enum:
            - FULL_TIME
            - PART_TIME
            - OTHER
          example: FULL_TIME

        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - RESIGNED
            - ON_LEAVE
          example: ACTIVE

        assignedSubjects:
          type: array
          items:
            $ref: '#/components/schemas/AssignedSubjectItem'
    FacultyPayload:
      type: object
      required:
        - employeeId
        - firstName
        - lastName
        - position
        - employmentType
      properties:
        employeeId:
          type: string
          example: EMP-001

        firstName:
          type: string
          example: Juan

        lastName:
          type: string
          example: Dela Cruz

        middleName:
          type: string
          nullable: true
          example: Santos

        email:
          type: string
          format: email
          example: juan.delacruz@example.com

        position:
          type: string
          enum:
            - FACULTY
            - PROGRAM_HEAD
            - DEAN

        employmentType:
          type: string
          enum:
            - FULL_TIME
            - PART_TIME
            - OTHER

        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - RESIGNED
            - ON_LEAVE
...
    AssignedSubjectItem:
      type: object
      required:
        - assignmentId
        - facultyId
        - subjectCode
        - sectionId
        - semester
        - academicYear
      properties:
        assignmentId:
          type: string
          example: ASG-001

        facultyId:
          type: string
          example: FAC-001

        subjectCode:
          type: string
          example: IT101

        subjectName:
          type: string
          example: Introduction to Information Technology

        sectionId:
          type: string
          example: SEC-001

        sectionName:
          type: string
          example: BSIT-3A

        semester:
          type: string
          enum:
            - FIRST
            - SECOND
            - SUMMER
          example: FIRST

        academicYear:
          type: string
          example: 2026-2027
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Resource not found

        error:
          type: string
          nullable: true
          example: NOT_FOUND

        details:
          type: object
          additionalProperties: true
          nullable: true

  # ==========================================================
  # COMMON RESPONSES
  # ==========================================================
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: User does not have permission to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'